requirements.txt
click==8.0.4
contourpy==1.0.5
cycler==0.11.0
fonttools==4.25.0
h5py==3.7.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
nltk==3.7
numpy==1.24.3
packaging==23.0
Pillow==9.4.0
pip==23.1.2
ply==3.11
pyparsing==3.0.9
PyQt5-sip==12.11.0
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.7.9
setuptools==67.8.0
sip==6.6.2
six==1.16.0
toml==0.10.2
torch==1.13.1
tornado==6.2
tqdm==4.65.0
typing_extensions==4.6.3
wheel==0.38.4
